pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guardian-drive"
version = "0.1.0"
description = "Multimodal driver impairment risk detection (physio + in-cabin video) with robustness, latency, and deployment discipline."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Ak"}]

dependencies = [
  "numpy>=1.26",
  "scipy>=1.11",
  "pandas>=2.1",
  "scikit-learn>=1.4",
  "torch>=2.2",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

